requirements.txt
asgiref==3.4.1
Django==3.2.9
django-ckeditor==6.2.0
django-js-asset==1.2.2
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.0
Unipath==1.1